change UserSongRating value to BigDecimal type

+++ src/main/java/com/lazydev/stksongbook/webapp/service/UserSongRatingService.java
@@ -8,6 +8,7 @@ import com.lazydev.stksongbook.webapp.service.exception.ForbiddenOperationExcept
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -39,15 +40,15 @@ public class UserSongRatingService {
     return repository.findByUserId(id);
   }
 
-  public List<UserSongRating> findByRating(Double rating) {
+  public List<UserSongRating> findByRating(BigDecimal rating) {
     return repository.findByRating(rating);
   }
 
-  public List<UserSongRating> findByRatingGreaterThanEqual(Double rating) {
+  public List<UserSongRating> findByRatingGreaterThanEqual(BigDecimal rating) {
     return repository.findByRatingGreaterThanEqual(rating);
   }
 
-  public List<UserSongRating> findByRatingLessThanEqual(Double rating) {
+  public List<UserSongRating> findByRatingLessThanEqual(BigDecimal rating) {
     return repository.findByRatingLessThanEqual(rating);
   }
